Split g4 file for db-discovery
